cockpit updates: drop the security branch
Based on upstream changes in commit b62d832 with comment...

This branch is very fragile, because the test won't fail if we
miss the match on the security needle. So in practice, we are
never going to notice when the needle goes stale, and we'll just
wind up never triggering this branch and always going down the
other path. That's the current situation: the security_install
needle last matched more than a year ago at least. Let's just
admit the truth here and drop the branch entirely.

Also update the cockpit_updates_restart_ignore needle. This is
in a similar case - we don't really notice when it goes stale,
as the test completes, it just takes a bit longer - but since
this one is quite easy to find, let's just update it instead of
dropping it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>

...additionally remove the rocky- needles created to replace the upstream needles.